chore: :art: Reorganize folders structure
