--- lib/media.js
@@ -1,8 +1,0 @@
-import { getStrapiURL } from "./api";
-
-export function getStrapiMedia(media) {
-  const imageUrl = media.url.startsWith("/")
-    ? getStrapiURL(media.url.substring(1))
-    : media.url;
-  return imageUrl;
-}
