+++ pages/contact.jsx
@@ -1,6 +1,6 @@
 import { Text } from "dana-react";
 
-import Layout from "../components/Layout";
+import Layout from "../components/layout/Layout";
 
 export default function Contact() {
 	return (
